==> sequence_to_sequence/models_with_clusters/1. build_complete_vocab.ipynb <==
import os, os.path,sys
import csv
import collections
import subprocess
import pandas as pd

from tqdm import tqdm
csv.field_size_limit(sys.maxsize)

!pwd

os.chdir("/data_data/session_length/nishanth01")

!pwd

!head data/tracks/tracks.csv

tracks_file = '/data_data/session_length/nishanth01/data/tracks/tracks.csv'
tracks_vocab_file = '/data_data/session_length/nishanth01/data/tracks_vocab.csv'
artist_vocab_file = '/data_data/session_length/nishanth01/data/artist_vocab.csv'

try:
    os.remove(tracks_vocab_file) 
    os.remove(artist_vocab_file) 
except OSError:
    pass


tracks_counter = collections.Counter()
artist_counter = collections.Counter()
with open(tracks_file) as tsvin:
    tsvin = csv.reader(tsvin)
    i = 0
    for row in tsvin:
        try:
            artist = row[1]
            track = row[2]
            if(artist):
                artist_counter.update([artist.lower().strip()])
            if(track):
                tracks_counter.update([track.lower().strip()])
#             if(i==1000):
#                 break
        except Exception:
            print('EXCEPTION: ',row)
            pass

        i = i+1

print ("Processing tracks vocab file...")  
with open(tracks_vocab_file, 'w') as writer:
    writer = csv.writer(writer,quoting=csv.QUOTE_NONNUMERIC,delimiter='\t')
    for word, count in tracks_counter.most_common():
        row = [word]
        row.append(count)
        writer.writerow(row)

        
print ("Processing artist vocab file...")  
with open(artist_vocab_file, 'w') as writer:
    writer = csv.writer(writer,quoting=csv.QUOTE_NONNUMERIC,delimiter='\t')
    for word, count in artist_counter.most_common():
        row = [word]
        row.append(count)
        writer.writerow(row)

print('Processed {0} records.'.format(i))    

!head $artist_vocab_file 

!head $tracks_vocab_file
