==> sequence_to_sequence/models_with_clusters/6. train_and_test_model.ipynb <==
import util
from collections import namedtuple

def get_params():
    hps_dict = {
        'data_file' : '/home/ubuntu/data/user_profile_cluster.csv',
        'use_spectral_clustering':True,
        'cluster_columns':(0,8),
        'clusters':5,
        'standardize':True,
        'model_dir':'/home/ubuntu/data/models/',
        'train_file':'/home/ubuntu/data/summary/train/{0}.csv',
        'test_file':'/home/ubuntu/data/summary/test/{0}.csv',
        'validate_file':'/home/ubuntu/data/summary/validate/{0}.csv',
        'baseline_mae':5429.204,#4828.088,
        'baseline_mae_std':0.754,
    }
    
    hps = namedtuple("HParams", hps_dict.keys())(**hps_dict)
    return hps

util.get_baseline_mae(get_params())

%%writefile testModel2.py

import pandas as pd
from collections import namedtuple
import util


def get_params():
    hps_dict = {
        'data_file' : '/home/ubuntu/data/user_profile_cluster.csv',
        'use_spectral_clustering':True,
        'cluster_columns':(0,8),
        'clusters':5,
        'standardize':False,
        'mix_std':True,
        'train_file':'/home/ubuntu/data/summary/train/{0}.csv',
        'test_file':'/home/ubuntu/data/summary/test/{0}.csv',
        'validate_file':'/home/ubuntu/data/summary/validate/{0}.csv',
        'model_dir':'/home/ubuntu/data/models/',
        'loss_func':'mae',
        'optimizer':'adam',
        'hidden_dim' : 250,
        'batch_size' : 400,
        'epochs' : 500,
        'layered':False,
        'layer_dims':150,
        'no_layers':1,
        'model_lstm':True,
        'filter_outliers':False,
        'baseline_mae':4932.859,#4828.088,
        'baseline_mae_std':0.766,
        'dropout':[0.0],
        'verbose_level':2,
        'plot_eval':False
    }
    
    hps = namedtuple("HParams", hps_dict.keys())(**hps_dict)
    return hps

def run():
    hps = get_params()
    users = util.train(hps)    
    util.test(users,hps)
        
        
if __name__ == '__main__':
    run()

!chmod 777 testModel2.py

#!python testModel2.py

!ls

#!rm nohup.out

!tail -30 nohup.out

!ps -ef | grep "testModel2.py"

#!kill -9 2790
